Decreased runtime.
Previously I had a for loop in which I would have to fully access the oldID
from the HDF5 file.  Changed it so it loads all the oldIDs into memory and then
just indexes them itself.

I'm sorry for doubting your words Mr Manodeep!

+++ genesis/utils/common.py
@@ -141,7 +141,7 @@ def temporalID_to_snapnum(temporalID, index_mult_factor):
         to the temporal ID.
     """
 
-    snapnum = int((temporalID - 1) / index_mult_factor)
+    snapnum = ((temporalID - 1) / index_mult_factor).astype(int)
 
     return snapnum
